+++ cmd/lino/dep_dataconnector.go
@@ -28,6 +28,8 @@ func dataconnectorStorage() domain.Storage {
 
 func dataPingerFactory() map[string]domain.DataPingerFactory {
 	return map[string]domain.DataPingerFactory{
-		"postgres": infra.NewSQLDataPingerFactory(logger),
+		"postgres":   infra.NewSQLDataPingerFactory(logger),
+		"godror":     infra.NewSQLDataPingerFactory(logger),
+		"godror-raw": infra.NewSQLDataPingerFactory(logger),
 	}
 }